requirements.txt
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
email-validator==2.0.0.post2 